requirements.txt
yfinance
pandas
sqlalchemy
psycopg2-binary
streamlit